% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger_level.R
\name{.build_signature}
\alias{.build_signature}
\title{Build API signature for logging to 'Azure Log Analytics'}
\usage{
.build_signature(
  customer_id,
  shared_key,
  date,
  content_length,
  method,
  content_type,
  resource
)
}
\arguments{
\item{customer_id}{\code{customer_id} of the 'Azure Log Analytics' workspace}

\item{shared_key}{\code{shared_key} of the 'Azure Log Analytics' workspace}

\item{date}{date-time of logging event}

\item{content_length}{Content length of the body}

\item{method}{Only one value is expected - \code{POST}}

\item{content_type}{Only one value is expected - \code{application/json}}

\item{resource}{Only one value is expected - \code{/api/logs}}
}
\value{
Returns part of the header of HTTP POST request to be sent to 'Azure
Log Analytics' workspace
}
\description{
'Azure Log Analytics' HTTP REST API documentation for 'Python' is followed to
create the 'R' version of it. 'Python' version of this function is described
at \url{https://learn.microsoft.com/en-us/azure/azure-monitor/logs/data-collector-api?tabs=python#sample-requests/}
}
